requirements.txt
streamlit
python-dotenv
google-search-results
langchain
langchain-openai
openai